# Docs/openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: CyberTrust KSA Platform API
  version: 1.0.0
  description: |
    Versioned API contracts for reusable policy evaluation, common-control mappings,
    connector lifecycle, automated control testing, events, and governed evidence decisions.
    All platform endpoints except health require a bearer JWT. Production service-to-service
    OAuth2 client credentials and signed webhook secrets are deployment requirements.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /platform/policy-evaluations/:
    post:
      summary: Evaluate an approved policy for the authenticated company tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy_version_id]
              properties:
                policy_version_id: {type: integer}
      responses:
        '201': {description: Reproducible evaluation with policy hash and reasons}
        '409': {description: Policy is not active and effective}
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT